Added executable file to the installation
And some testing data in `tests/data`
